+++ src/util.ts
@@ -1,0 +1,13 @@
+import { treeGraph } from './util/tree-graph.js';
+import exitWithError from './util/exit-with-error.js';
+import { countSubcommands } from './util/count-commands.js';
+import { navigateSubcommands } from './util/navigate-subcommands.js';
+import { resolveDynamic } from './util/resolve-dynamic.js';
+
+export {
+    exitWithError,
+    treeGraph,
+    countSubcommands,
+    navigateSubcommands,
+    resolveDynamic,
+};